lib/telemetry:fix telemetry conns leak in case of socket write fail
Telemetry can only create 10 conns by default, each of which is processed by a thread.
When a thread fails to write using socket, the thread will end directly without reducing the total number of conns.
This will result in the machine running for a long time, and if there are 10 failures, the telemetry will be unavailable

Signed-off-by: Sun shaowei <1819846787@qq.com>